Handle random interrupts on dial action pin when normal dialing starts.
Sometimes, unexpected interrupts happen on the pin that monitors
the start and end of a digit input process. If we detect a suspiciously
large timespan between a digit start interrupt and a digit end
interrupt, we now assume it’s a digit start rather than a digit end
to cope with the problem.